Update to Next.js v16
