+++ next.config.js
@@ -11,29 +11,16 @@
 
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  typescript: {
-    // !! WARN !!
-    // Dangerously allow production builds to successfully complete even if
-    // your project has type errors.
-    // !! WARN !!
-    ignoreBuildErrors: true,
-  },
-  eslint: {
-    // Warning: This allows production builds to successfully complete even if
-    // your project has ESLint errors.
-    ignoreDuringBuilds: true,
-  },
-  webpack: (config, { isServer }) => {
-    // Fix canvas dependency issues during build
-    if (isServer) {
-      config.externals = config.externals || [];
-      config.externals.push('canvas');
-    }
-    return config;
-  },
-}
+	typescript: {
+		// !! WARN !!
+		// Dangerously allow production builds to successfully complete even if
+		// your project has type errors.
+		// !! WARN !!
+		ignoreBuildErrors: true,
+	},
+};
 
 // STANDARD CONFIG (Workflow DevKit package removed due to beta issues)
 // The workflow functions in lib/workflows/ execute as regular async functions.
 // They work perfectly, just without auto-durability/retries until stable release.
-module.exports = nextConfig
+module.exports = nextConfig;
