+++ package.json
@@ -52,13 +52,13 @@
     "inngest": "^3.45.0",
     "input-otp": "1.4.1",
     "lucide-react": "^0.454.0",
-    "next": "14.0.4",
+    "next": "16.0.1",
     "next-themes": "^0.4.4",
     "openai": "^6.8.1",
     "pdf-parse": "^2.4.5",
-    "react": "^18.2.0",
+    "react": "19.2.0",
     "react-day-picker": "8.10.1",
-    "react-dom": "^18.2.0",
+    "react-dom": "19.2.0",
     "react-force-graph-2d": "^1.27.1",
     "react-hook-form": "^7.54.1",
     "react-hot-toast": "^2.4.1",
@@ -72,13 +72,18 @@
   },
   "devDependencies": {
     "@types/node": "^20.10.4",
-    "@types/react": "^18.2.45",
-    "@types/react-dom": "^18.2.18",
+    "@types/react": "19.2.2",
+    "@types/react-dom": "19.2.2",
     "autoprefixer": "^10.4.16",
-    "eslint": "^8.56.0",
-    "eslint-config-next": "^14.0.4",
+    "eslint": "^9.39.1",
+    "eslint-config-next": "16.0.1",
     "postcss": "^8.4.32",
     "tailwindcss": "^3.4.17",
     "typescript": "^5.3.3"
-  }
+  },
+  "overrides": {
+    "@types/react": "19.2.2",
+    "@types/react-dom": "19.2.2"
+  },
+  "packageManager": "pnpm@9.15.4+sha512.b2dc20e2fc72b3e18848459b37359a32064663e5627a51e4c74b2c29dd8e8e0491483c3abb40789cfd578bf362fb6ba8261b05f0387d76792ed6e23ea3b1b6a0"
 }
